Initial commit.
Includes the general design of the app and api that I'd like it to be.
Needs a lot of tidyup, especially on the json sent to the front end.

Lots of minor features to add like swipe support, pagination.
If I can apply for elevated, use V1 instead of V2 for media searching.
V2 is not ideal for pulling media only.

Future features would be a database for "caching" the information.
This would help keep TwitterAPI calls down and speed up retrieval.