JSON Reduction, tidy-up and minor tweaks
At the moment, navigating to a handle or refreshing an active handle page will not show the search bar until the page is loaded.

I'll have to make a nav component for the next update to separate out the html, this should already be a thing but ya'know.
So until then, pages will load weirdly if navigated to by the browser bar rather than using the app "normally"

- Filter buttons now flip so one is always selected
- JSON from back->frontend has been significantly reduced
- URLs are now prebuilt on the backend to reduce frontend workload
- Error messages are now passed via the backend
- Removed user deduction step on frontend, it's now a part of backend