Pagination workaround, page loading fix.
Twitter V2 API was bizarre with pagination tokens, where sometimes they would not return as expected.
To fix this, I've swapped it out for the oldest tweet ID in the results, and then checked the future to see if there are remaining results.

This isn't ideal, but it works up until the next soft-limit in the API. So some accounts will still bottom out early before all media is displayed.
To make this apparent, a message is displayed to the user when no more results are available.

Other than this, I adjusted the angular resolver to not do stupid things like call the API, I was tired when I wrote it.
The page now loads correctly and logically, with the little spinny thing too c: